Installer/*: Similar adjustments for VBOX_WITH_WEBSERVICES. Fix for solaris typo two commits back. bugref:10030
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@98681 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Installer/linux/install.sh
@@ -339,7 +339,9 @@ if [ "$ACTION" = "install" ]; then
         ln -sf $INSTALLATION_DIR/VBox.sh /usr/bin/VBoxBugReport
     fi
     ln -sf $INSTALLATION_DIR/VBox.sh /usr/bin/VBoxAutostart
-    ln -sf $INSTALLATION_DIR/VBox.sh /usr/bin/vboxwebsrv
+    if [ -f $INSTALLATION_DIR/vboxwebsrv ]; then
+        ln -sf $INSTALLATION_DIR/VBox.sh /usr/bin/vboxwebsrv
+    fi
     ln -sf $INSTALLATION_DIR/vbox-img /usr/bin/vbox-img
     ln -sf $INSTALLATION_DIR/vboximg-mount /usr/bin/vboximg-mount
     if [ -d /usr/share/pixmaps/ ]; then
